Rename integration test mockApis to more closely match the api name
