+++ cypress.config.ts
@@ -3,14 +3,13 @@ import cypressMochawesomeReportPlugin from 'cypress-mochawesome-reporter/plugin'
 import fs from 'fs'
 
 import auth from './integration_tests/mockApis/auth'
-import caseloads from './integration_tests/mockApis/caseloads'
 import courses from './integration_tests/mockApis/courses'
 import courseParticipations from './integration_tests/mockApis/courseParticipations'
 import manageUsers from './integration_tests/mockApis/manageUsers'
-import prisons from './integration_tests/mockApis/prisons'
-import prisoners from './integration_tests/mockApis/prisoners'
+import prisonApi from './integration_tests/mockApis/prison'
+import prisonRegisterApi from './integration_tests/mockApis/prisonRegister'
+import prisonerSearch from './integration_tests/mockApis/prisonerSearch'
 import referrals from './integration_tests/mockApis/referrals'
-import sentenceInformation from './integration_tests/mockApis/sentenceInformation'
 import tokenVerification from './integration_tests/mockApis/tokenVerification'
 import { resetStubs } from './wiremock'
 
@@ -36,14 +35,13 @@ export const defaultConfig: Cypress.ConfigOptions = {
       on('task', {
         reset: resetStubs,
         ...auth,
-        ...caseloads,
         ...courses,
         ...courseParticipations,
         ...manageUsers,
-        ...prisons,
-        ...prisoners,
+        ...prisonApi,
+        ...prisonRegisterApi,
+        ...prisonerSearch,
         ...referrals,
-        ...sentenceInformation,
         ...tokenVerification,
         log(message) {
           console.log(message)
